GitIgnore Update
Brought this from all of the other projects.

Listening to "Haunted", my new song from the other day. Very happy with how it sounds.

It was very foggy at the lake today, it really felt great for my vibe today. Kind of "off", great description. Needed to get out of the house and run.