XDecompress Rust Port
Getting help from GPT, going to look over how this works after getting some insight on how the existing code functions. It's still just a little over my head still.

Listening to "From the Depths of Subconscious". It is part of a project I started working on a month of two ago, where I wanted it to sound and feel like one of my early projects, but with the new mixing and mastering techniques. Excited for when that one is complete and released, as well as just getting to listen to it again. I want to still continue making stuff like that again. It's interesting to figure out the common thread of what your inspirations were at the time, then come back to that, and write similar things with similar goals, they really can come from the same place in a way. It's never going to be the exact same frame of mind, since time has passed, things have been learned, forgotten, all of it. But you can still embody what that used to be, even if it's not the "you" it once was. Even if "you now" is something or someone different, you can still connect to the old you :)