Supposed 360 Parsing
Listening to Ocean Machine. Wow, I really do love this album. To be totally open, I think since I did Devin's catalog backwards, this once originally felt derivative to me. But it was actually the other way around. I already knew that, but now I feel it. This is where the heart of it all started. It's fantastic.

Will see if this code works, going to trim some things away, code that's unused. This module is simply to demo whether my LZX wrapper/parser is actually working as intended, I think the Dat for 360 saves isn't plain LZX, hence why it wasn't working initially. And I want to make sure it's working in it's current state before trying to migrate it all to Rust.