Error Flattening
I like using the `map_err` over the match syntax, when applicable of course. To me, I feel like it cements that there are no side-effects, and that the statement is "complete" so to say, as well. If there were more things to add, and maybe it wasn't done yet, I might leave the match block there for future changes.