Simple Function Formatting
Heck yeah, love this! Very clear as to what this does now. Hopefully it will run, lol. Always happy to debug.

If anything, this is just helping me learn how to convert C/C++ to Rust! That's the ideal goal here too. If I can do that, then it doesn't matter in the end, I can just keep trying to rewrite the logic, simply by understanding how the code works in the first place!!!! "The ultimate GigaChad move®" I just made that up