Debug the Correct File
This was from whenever in the last few days, I didn't get to commit this. I was starting to debug the outputs, and realized I was opening the wrong file since having implemented for the `.dat` file, which is what has the compression I am trying to open. Will come back to this again once it is more apparent what I was going to try next.